abstracted stripe components, brought in redux to Billing component
